requirement.txt
delta-spark==1.0.0
importlib-metadata==4.8.1
py4j==0.10.9
pyspark==3.1.2
zipp==3.6.0
